Added Uploads Directory
During the initial upload of the program's code, the Uploads directory didn't upload to the repository, so I added an empty CSV file to the directory to upload it to the repository. Still, the fact that the repository didn't accept the Uploads depository initially exposes it as unnecessary, highlighting a need to update the program to stop using this repository in the future.